+++ CHANGELOG.md
@@ -5,7 +5,9 @@ v0.3 (TBD)
  - Catch tag edit exceptions, display error log in window
  - Reorganized code so it is easier to follow, lots of comments added
  - Search for DICOM tags
- - Values can be dynamically defined based on file_path or current tag values  
+ - Values can be dynamically defined based on file_path or current tag values
+ - Thread file parsing with progress bar
+ 
  
 v0.2 (2020.03.27)
 --------------------